doc_denoiser/requirements.txt
# Requirements for Document Denoiser
# Install with: pip install -r requirements.txt

torch
torchvision
streamlit
Pillow
numpy
opencv-python
PyMuPDF
